edit fake data,rename get course details function
