+++ server/src/controllers/getCourseDetails.js
