+++ server/src/controllers/index.js
@@ -1,7 +1,7 @@
 const { getCoursesByCatId } = require('./middleware/getCatcourses');
 
 const { clientError, serverError } = require('./errorHandle');
-const { getCourseById } = require('./course');
+const { getCourseById } = require('./getCourseDetails');
 
 module.exports = {
   clientError,
