+++ server/src/database/config/fakeData.sql
@@ -4,7 +4,7 @@ VALUES(
     '1',
     'computer',
     'Moh',
-    'www.course1.com',
+    'www.course1.com/1',
     '3.4',
     '23 reviews',
     'Hi from computer course',
@@ -14,7 +14,7 @@ VALUES(
     '2',
     'Science',
     'Ralef',
-    'www.course1.com',
+    'www.course1.com/2',
     '4.4',
     '60 reviews',
     'Hi from Science course',
@@ -23,7 +23,7 @@ VALUES(
 (   '3',
     'Phonetics',
     'Yaser',
-    'www.course1.com',
+    'www.course1.com/3',
     '5.4',
     '50 reviews',
     'Hi from Phonetics course',
@@ -33,7 +33,7 @@ VALUES(
     '4',
     'Linguistics',
     'Moh',
-    'www.course1.com',
+    'www.course1.com/4',
     '1.3',
     '80 reviews',
     'Hi from Linguistics course',
